+++ src/helpers/colors/create-cursor-color.ts
@@ -1,0 +1,17 @@
+import { type CursorColor, type PaletteColors } from '../../types';
+
+export function createCursorColor({
+  cursorColor,
+  paletteColors,
+}: {
+  cursorColor: CursorColor;
+  paletteColors: PaletteColors;
+}) {
+  const cursorTyped = cursorColor as keyof typeof paletteColors;
+  const color = paletteColors[cursorTyped] ?? paletteColors.fg;
+
+  return {
+    'editorCursor.foreground': color,
+    'terminalCursor.foreground': color,
+  };
+}