+++ src/helpers/colors/create-selection-colors.ts
@@ -8,19 +8,27 @@ import {
 export function createSelectionColors({
   contrastColors,
   paletteColors,
-  selection,
+  selectionColor,
 }: {
   contrastColors: ContrastColors;
   paletteColors: PaletteColors;
-  selection: SelectionColor;
-}): SelectionColors {
-  const isGrey = selection === 'grey';
-  const dimColor = `${selection}Dim` as keyof typeof paletteColors;
+  selectionColor: SelectionColor;
+}) {
+  const isGrey = selectionColor === 'grey';
+  const dimColor = `${selectionColor}Dim` as keyof typeof paletteColors;
   const paletteColor = paletteColors[dimColor] ?? contrastColors.bg6;
 
-  return {
+  const colors: SelectionColors = {
     bg1: `${paletteColor}${isGrey ? 'd0' : '60'}`,
     bg2: `${paletteColor}${isGrey ? 'b0' : '40'}`,
     bg3: `${paletteColor}${isGrey ? '58' : '20'}`,
+  };
+
+  return {
+    'editor.findRangeHighlightBackground': colors.bg3,
+    'editor.inactiveSelectionBackground': colors.bg3,
+    'editor.selectionBackground': colors.bg2,
+    'editor.selectionHighlightBackground': colors.bg3,
+    'selection.background': colors.bg1,
   };
 }