+++ src/helpers/colors/index.ts
@@ -1,4 +1,5 @@
-export * from './create-colors';
+export * from './create-cursor-color';
+export * from './create-editor-colors';
 export * from './create-selection-colors';
 export * from './create-semantic-colors';
 export * from './create-separators-color';